requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
psutil>=5.9.0
aiofiles>=23.0.0
pydantic==2.6.0
pydantic-settings==2.2.0 